refactor pages to components
